+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 ### Fixed
 
 - Forward delete selection now deletes selection.
+- Only show fonts that support the script of the currently selected languages.
 
 ### Maintenance